Address PR review: atomicity, pool drift, negative budget, inline import
- Extract _consume_budget into rumil.budget so rumil.calls.page_creators
  can import it at the top of the file rather than inside a function. The
  inline import was a circular-import workaround
  (orchestrators.common -> calls.find_considerations -> calls.page_creators);
  the new module breaks the cycle. Existing import sites in
  orchestrators.common and views.sectioned now point at the new location;
  orchestrators.common keeps a re-export for any external callers that
  imported from there.

- Recurse dispatches now use the atomic qbp_recurse RPC instead of
  qbp_consume(parent) + child.run()'s qbp_register(child). The two-step
  flow left a window where peer cycles could see the parent debited but
  the child unregistered, and lost the budget entirely if child.run()
  raised before reaching its register. The new pool_pre_registered flag
  on the orchestrator constructors tells the child's run() to skip its
  own register (since qbp_recurse already did it) while still calling
  qbp_unregister in the finally block to balance active_calls.

- The main orchestrator loops' view.refresh() calls now pass
  pool_question_id=self.pool_question_id, matching the dispatch-sequence
  refresh path. Without this the per-question pool's `consumed` drifted
  under what was actually spent on the question, causing peers to over-
  spend.

- When the post-fresh-snapshot budget is <=0 (or <=1 in the experimental
  variant whose dispatch_budget = budget - 1), bail before calling the
  prio LLM rather than rendering "budget -1" verbatim into the prompt.
  Marks the prio call complete with a "pool drained by peer cycles"
  reason and returns an empty result so the outer loop breaks.
