+++ src/rumil/orchestrators/claim_investigation.py
@@ -68,6 +68,7 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
         db: DB,
         broadcaster: Broadcaster | None = None,
         budget_cap: int | None = None,
+        pool_pre_registered: bool = False,
     ):
         super().__init__(db, broadcaster)
         self._invocation: int = 0
@@ -80,6 +81,10 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
         self._parent_call_id: str | None = None
         self._sequence_id: str | None = None
         self._seq_position: int = 0
+        # See TwoPhaseOrchestrator for why this exists. When True, the
+        # parent already registered our contribution via qbp_recurse, so
+        # run() must not double-register.
+        self._pool_pre_registered: bool = pool_pre_registered
 
     def _effective_budget(self, global_remaining: int) -> int:
         if self._budget_cap is not None:
@@ -132,8 +137,9 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
             self._sequence_id = seq.id
             self._seq_position = 0
         self.pool_question_id = claim_id
-        contribution = self._budget_cap if self._budget_cap is not None else effective
-        await self.db.qbp_register(claim_id, contribution)
+        if not self._pool_pre_registered:
+            contribution = self._budget_cap if self._budget_cap is not None else effective
+            await self.db.qbp_register(claim_id, contribution)
         try:
             while True:
                 remaining = await self.db.budget_remaining()
@@ -205,6 +211,7 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
                         force=True,
                         sequence_id=self._sequence_id,
                         sequence_position=self._seq_position,
+                        pool_question_id=self.pool_question_id,
                     )
                     if self._sequence_id is not None:
                         self._seq_position += 1
@@ -547,6 +554,15 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
         # context section agree on what's available.
         fresh_pool = await self.db.qbp_get(claim_id)
         budget = min(budget, max(fresh_pool.remaining, 0))
+        # If the pool drained between top-of-loop and now (i.e. peer cycles
+        # consumed our slice), bail before calling the LLM with budget 0.
+        if budget <= 0:
+            await mark_call_completed(
+                p_call,
+                self.db,
+                "Pool drained by peer cycles before this round could plan.",
+            )
+            return PrioritizationResult(dispatch_sequences=[], call_id=p_call.id)
         context_text, short_id_map = await build_prioritization_context(
             self.db,
             scope_question_id=claim_id,
@@ -609,11 +625,12 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
                         d.payload.question_id[:8],
                     )
                     continue
-                await self.db.qbp_consume(claim_id, d.payload.budget)
+                await self.db.qbp_recurse(claim_id, resolved, d.payload.budget)
                 child = ClaimInvestigationOrchestrator(
                     self.db,
                     self.broadcaster,
                     budget_cap=d.payload.budget,
+                    pool_pre_registered=True,
                 )
                 child._parent_call_id = p_call.id
                 children.append((child, resolved))
@@ -631,11 +648,12 @@ class ClaimInvestigationOrchestrator(BaseOrchestrator):
                         d.payload.question_id[:8],
                     )
                     continue
-                await self.db.qbp_consume(claim_id, d.payload.budget)
+                await self.db.qbp_recurse(claim_id, resolved, d.payload.budget)
                 child = TwoPhaseOrchestrator(
                     self.db,
                     self.broadcaster,
                     budget_cap=d.payload.budget,
+                    pool_pre_registered=True,
                 )
                 child._parent_call_id = p_call.id
                 children.append((child, resolved))
