+++ src/rumil/orchestrators/experimental.py
@@ -73,6 +73,7 @@ class ExperimentalOrchestrator(BaseOrchestrator):
         db: DB,
         broadcaster: Broadcaster | None = None,
         budget_cap: int | None = None,
+        pool_pre_registered: bool = False,
     ):
         super().__init__(db, broadcaster)
         self._invocation: int = 0
@@ -86,6 +87,8 @@ class ExperimentalOrchestrator(BaseOrchestrator):
         self._sequence_id: str | None = None
         self._seq_position: int = 0
         self._last_linker_eval_at: datetime | None = None
+        # See TwoPhaseOrchestrator for why this exists.
+        self._pool_pre_registered: bool = pool_pre_registered
 
     def _effective_budget(self, global_remaining: int) -> int:
         if self._budget_cap is not None:
@@ -143,8 +146,9 @@ class ExperimentalOrchestrator(BaseOrchestrator):
             self._seq_position = 0
         budget_token = set_experimental_scout_budget(effective)
         self.pool_question_id = root_question_id
-        contribution = self._budget_cap if self._budget_cap is not None else effective
-        await self.db.qbp_register(root_question_id, contribution)
+        if not self._pool_pre_registered:
+            contribution = self._budget_cap if self._budget_cap is not None else effective
+            await self.db.qbp_register(root_question_id, contribution)
         try:
             while True:
                 remaining = await self.db.budget_remaining()
@@ -212,6 +216,7 @@ class ExperimentalOrchestrator(BaseOrchestrator):
                         force=True,
                         sequence_id=self._sequence_id,
                         sequence_position=self._seq_position,
+                        pool_question_id=self.pool_question_id,
                     )
                     if self._sequence_id is not None:
                         self._seq_position += 1
@@ -609,6 +614,15 @@ class ExperimentalOrchestrator(BaseOrchestrator):
         # context section agree on what's available.
         fresh_pool = await self.db.qbp_get(question_id)
         budget = min(budget, max(fresh_pool.remaining, 0))
+        # If the pool drained between top-of-loop and now (i.e. peer cycles
+        # consumed our slice), bail before calling the LLM with budget 0/-1.
+        if budget <= 1:
+            await mark_call_completed(
+                p_call,
+                self.db,
+                "Pool drained by peer cycles before this round could plan.",
+            )
+            return PrioritizationResult(dispatch_sequences=[], call_id=p_call.id)
         context_text, short_id_map = await build_prioritization_context(
             self.db,
             scope_question_id=question_id,
@@ -677,11 +691,12 @@ class ExperimentalOrchestrator(BaseOrchestrator):
                         d.payload.question_id[:8],
                     )
                     continue
-                await self.db.qbp_consume(question_id, d.payload.budget)
+                await self.db.qbp_recurse(question_id, resolved, d.payload.budget)
                 child = ExperimentalOrchestrator(
                     self.db,
                     self.broadcaster,
                     budget_cap=d.payload.budget,
+                    pool_pre_registered=True,
                 )
                 child._parent_call_id = p_call.id
                 children.append((child, resolved))
