+++ src/rumil/orchestrators/two_phase.py
@@ -70,6 +70,7 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
         db: DB,
         broadcaster: Broadcaster | None = None,
         budget_cap: int | None = None,
+        pool_pre_registered: bool = False,
     ):
         super().__init__(db, broadcaster)
         self._invocation: int = 0
@@ -82,6 +83,12 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
         self._parent_call_id: str | None = None
         self._sequence_id: str | None = None
         self._seq_position: int = 0
+        # When True, this orchestrator's contribution to the question pool
+        # was already registered atomically by the parent's qbp_recurse call,
+        # so run() must NOT register again (it would double-count). The
+        # finally block still calls qbp_unregister to balance active_calls
+        # (which qbp_recurse incremented via its register half).
+        self._pool_pre_registered: bool = pool_pre_registered
 
     def _effective_budget(self, global_remaining: int) -> int:
         if self._budget_cap is not None:
@@ -138,8 +145,9 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
             self._sequence_id = seq.id
             self._seq_position = 0
         self.pool_question_id = root_question_id
-        contribution = self._budget_cap if self._budget_cap is not None else effective
-        await self.db.qbp_register(root_question_id, contribution)
+        if not self._pool_pre_registered:
+            contribution = self._budget_cap if self._budget_cap is not None else effective
+            await self.db.qbp_register(root_question_id, contribution)
         try:
             while True:
                 remaining = await self.db.budget_remaining()
@@ -211,6 +219,7 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
                         force=True,
                         sequence_id=self._sequence_id,
                         sequence_position=self._seq_position,
+                        pool_question_id=self.pool_question_id,
                     )
                     if self._sequence_id is not None:
                         self._seq_position += 1
@@ -592,6 +601,16 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
         # context section agree on what's available.
         fresh_pool = await self.db.qbp_get(question_id)
         budget = min(budget, max(fresh_pool.remaining, 0))
+        # If the pool drained between top-of-loop and now (i.e. peer cycles
+        # consumed our slice), bail before calling the LLM with budget 0/-1.
+        # The outer loop will break on the empty result.
+        if budget <= 0:
+            await mark_call_completed(
+                p_call,
+                self.db,
+                "Pool drained by peer cycles before this round could plan.",
+            )
+            return PrioritizationResult(dispatch_sequences=[], call_id=p_call.id)
         context_text, short_id_map = await build_prioritization_context(
             self.db,
             scope_question_id=question_id,
@@ -675,11 +694,12 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
                         d.payload.question_id[:8],
                     )
                     continue
-                await self.db.qbp_consume(question_id, d.payload.budget)
+                await self.db.qbp_recurse(question_id, resolved, d.payload.budget)
                 child_claim = ClaimInvestigationOrchestrator(
                     self.db,
                     self.broadcaster,
                     budget_cap=d.payload.budget,
+                    pool_pre_registered=True,
                 )
                 child_claim._parent_call_id = p_call.id
                 children.append((child_claim, resolved))
@@ -697,11 +717,12 @@ class TwoPhaseOrchestrator(BaseOrchestrator):
                         d.payload.question_id[:8],
                     )
                     continue
-                await self.db.qbp_consume(question_id, d.payload.budget)
+                await self.db.qbp_recurse(question_id, resolved, d.payload.budget)
                 child = TwoPhaseOrchestrator(
                     self.db,
                     self.broadcaster,
                     budget_cap=d.payload.budget,
+                    pool_pre_registered=True,
                 )
                 child._parent_call_id = p_call.id
                 children.append((child, resolved))
